Updated CRCs for ST variant CCF/SCF tests with undocumented flags
Recent research has shown that the CCF/SCF instructions, on ST CMOS
variants, only change the 5th bit of the flag register (aka "Y") if
flags were modified by the previous instruction. If flags are stable,
then only bit 3 is set according to the A register.

This subtle difference in behavior does not appear to influence the
output of the z80ccfscr program, but it does change the register state
reflected in the checksums of the standard test suite.